+++ plotnine/geoms/geom_text.py
@@ -2,9 +2,17 @@ from contextlib import suppress
 
 from matplotlib.text import Text
 
+try:
+    from adjustText import adjust_text
+except ImportError:
+    HAS_ADJSUST_TEXT = False
+else:
+    HAS_ADJSUST_TEXT = True
+
 from ..utils import to_rgba
 from ..doctools import document
 from ..positions import position_nudge
+from ..exceptions import PlotnineError
 from .geom import geom
 
 
@@ -38,6 +46,11 @@ class geom_text(geom):
         Horizontal adjustment to apply to the text
     nudge_y : float (default: 0)
         Vertical adjustment to apply to the text
+    adjust_text: dict (default: None)
+        Parameters to :class:`adjustText.adjust_text` will repel
+        overlapping texts. This parameter takes priority of over
+        ``nudge_x`` and ``nudge_y``.
+        See https://github.com/Phlya/adjustText/wiki .
     format_string : str (default: None)
         If not :py:`None`, then the text if formatted with this
         string using :meth:`str.format`
@@ -55,16 +68,24 @@ class geom_text(geom):
                       'family': None, 'fontweight': 'normal',
                       'fontstyle': 'normal', 'ha': 'center',
                       'va': 'center', 'nudge_x': 0, 'nudge_y': 0,
+                      'adjust_text': None,
                       'format_string': None}
 
     def __init__(self, *args, **kwargs):
         nudge_kwargs = {}
-        with suppress(KeyError):
-            nudge_kwargs['x'] = kwargs['nudge_x']
-        with suppress(KeyError):
-            nudge_kwargs['y'] = kwargs['nudge_y']
-        if nudge_kwargs:
-            kwargs['position'] = position_nudge(**nudge_kwargs)
+        adjust_text = kwargs.get('adjust_text', None)
+        if adjust_text is None:
+            with suppress(KeyError):
+                nudge_kwargs['x'] = kwargs['nudge_x']
+            with suppress(KeyError):
+                nudge_kwargs['y'] = kwargs['nudge_y']
+            if nudge_kwargs:
+                kwargs['position'] = position_nudge(**nudge_kwargs)
+        elif not HAS_ADJSUST_TEXT:
+            raise PlotnineError(
+                "To use adjust_text you must install the adjustText "
+                "package."
+            )
 
         # Accomodate for the old names
         if 'hjust' in kwargs:
@@ -88,6 +109,12 @@ class geom_text(geom):
             data['label'] = ['${}$'.format(l) for l in data['label']]
 
         return data
+
+    def draw_panel(self, data, panel_params, coord, ax, **params):
+        super().draw_panel(data, panel_params, coord, ax, **params)
+        if params['adjust_text'] is not None:
+            texts = [t for t in ax.texts]
+            adjust_text(texts, ax=ax, **params['adjust_text'])
 
     @staticmethod
     def draw_group(data, panel_params, coord, ax, **params):