+++ plotnine/tests/test_geom_text_label.py
@@ -1,8 +1,9 @@
 import pandas as pd
 import numpy as np
 
-from plotnine import (ggplot, aes, geom_text, geom_label,
+from plotnine import (ggplot, aes, geom_text, geom_label, geom_point,
                       scale_size_continuous, scale_y_continuous)
+from plotnine.data import mtcars
 
 n = 5
 labels = ['ggplot', 'aesthetics', 'data', 'geoms',
@@ -12,9 +13,16 @@ df = pd.DataFrame({
         'y': range(n),
         'label': labels[:n],
         'z': range(n),
-        'angle': np.linspace(0, 90, n),
-
+        'angle': np.linspace(0, 90, n)
     })
+
+adjust_text = {
+    'expand_points': (2, 2),
+    'arrowprops': {
+        'arrowstyle': '->',
+        'color': 'red'
+    }
+}
 
 
 def test_text_aesthetics():
@@ -50,3 +58,19 @@ def test_label_aesthetics():
          scale_y_continuous(limits=(-0.5, n-0.5)))
 
     assert p == 'label_aesthetics'
+
+
+def test_adjust_text():
+    p = (ggplot(mtcars.tail(2), aes('mpg', 'disp', label='name'))
+         + geom_point(size=5, fill='black')
+         + geom_text(adjust_text=adjust_text)
+         )
+    assert p == 'adjust_text'
+
+
+def test_adjust_label():
+    p = (ggplot(mtcars.tail(2), aes('mpg', 'disp', label='name'))
+         + geom_point(size=5, fill='black')
+         + geom_label(adjust_text=adjust_text)
+         )
+    assert p == 'adjust_label'